chore: Updated checklist
chore: Reorganized the modules and tests
